feat(player-list): migrates to material ui
Migrates the component to use Material UI instead of Chakra UI

re #144
